Logging seems sorted for now
